requirements.txt
fastapi==0.115.12
matplotlib==3.8.4
Pillow==11.2.1
pydantic==2.11.3
Requests==2.32.3
tensorflow==2.18.0
uvicorn==0.34.1
python-multipart==0.0.12
